feat: 주문 기능 추가 feat: 포인트 사용 로그 기능 추가 refactor: 포인트 충전 Request, Response 네이밍 직관적 변경 (add -> charge) remove: 다수 테이블 기능 구현을 위해 삭제 및 변경 refactor: 금액, 수량 관련 long 구현 에서 BigInteger 로 변경

+++ src/main/java/kyjtheyj/coffeedrink/domain/menu/model/request/MenuRegisterRequest.java
@@ -24,6 +24,6 @@ public record MenuRegisterRequest(
 
         , @NotNull(message = "재고는 필수입니다")
         @Min(value = 0, message = "재고는 0개 이상이어야 합니다")
-        Long quantity
+        BigInteger quantity
 ) {
 }
